show checked in status without login
